Updated Recipe constructor and added user interaction
Updated the `Recipe` class constructor in `Recipe.cs` to take a string parameter `name` for setting the `RecipeName` property.

Added a new method `BeginHere()` to the `Program` class in `Program.cs`. This method interacts with the user, prompting them to enter a recipe name and creating a new `Recipe` object with the entered name.

The `Main` method of the `Program` class has been updated to create an instance of `Program` and call the `BeginHere()` method, initiating user interaction when the program is run.