Update calorie handling to per-unit basis
Updated the handling of calories in the application to be on a per-unit basis. Changes include updating the string representation of an ingredient in `Ingredient.cs` to specify that the calorie count is "per unit". In `Recipe.cs`, the calculation of total calories in a recipe now multiplies the calories per unit by the quantity of the ingredient. Also, the scaling and reset scaling functions in `Recipe.cs` no longer affect the calories of the ingredient. Lastly, in `Program.cs`, the user is now prompted to enter the number of calories per unit of measurement for an ingredient.