Initiated GUI development
This commit marks a significant advancement in the project by initiating the development of a graphical user interface (GUI) component and updating various configurations to support this development.

Key changes include the restructuring of the namespace in Program.cs to better align with project standards, and the addition of a new GUI project to the solution.

Additionally, the project now incorporates the MVVM pattern through the creation of ObservableObject.cs and RelayCommand.cs, facilitating a structured approach to UI development.

Visual customization and user experience improvements are evident in the addition of XAML resource dictionary files, MenuButtonTheme.xaml and TextBoxTheme.xaml, which define custom styles for UI elements, contributing to a more engaging and visually appealing interface.

These changes collectively underscore the project's evolution towards adopting modern .NET development practices and enhancing user interaction.