UI redesign for improved responsiveness
Implemented a significant UI redesign across the application to enhance responsiveness and modernize the user interface. In RecipesView.xaml, transitioned from a column-based to a row-based layout to better support a vertical layout, including adjustments to the Border element and the replacement of a vertical StackPanel with a horizontal one inside a ScrollViewer for filter and search components. This redesign aims for a more streamlined and consistent UI, introducing a new filter section with horizontal scrolling and replacing old filter UI elements with Grid, TextBlock, Button, and custom controls like xctk:CheckComboBox and xctk:RangeSlider.

In MainWindow.xaml, made minor adjustments to window size properties to align with the new design elements in RecipesView.xaml.

Enhanced SearchBoxTheme.xaml with the addition of an XML namespace for Xceed Toolkit controls and introduced new styles for TextBox, xctk:CheckComboBox, and Button with custom ControlTemplates, aiming for a unified and modern design across the application. This includes the use of a custom font, Poppins, for a polished UI appearance.

Overall, these changes aim to provide a more responsive, modern, and consistent user experience across the application.